% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aro.clo.points.r
\name{aro.clo.points}
\alias{aro.clo.points}
\title{aro.clo.points}
\usage{
aro.clo.points(vertex, landmarks)
}
\arguments{
\item{vertex}{matrix mesh vertex}

\item{landmarks}{numeric: a kxm matrix of a landmark set}
}
\value{
position numeric: a vector with vertex number nearest the landmark set
}
\description{
This function finds the vertices numbering nearest a landmark set
}
\examples{
\dontrun{
#load an example: mesh, and L set
data(exp.teeth.mesh)
data(exp.teeth.Lset)
sur=t(vcgQEdecim(teeth.mesh,percent=0.50)$vb[-4, ])
set=teeth.Lset
example=aro.clo.points(vertex=sur,landmarks=set)
}
}
\author{
Antonio Profico, Alessio Veneziano, Alessandro Lanteri
}

