% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bezier.amira.path.r
\name{bezier.amira.path}
\alias{bezier.amira.path}
\title{bezier.amira.path}
\usage{
bezier.amira.path(path.name, npoint, mag)
}
\arguments{
\item{path.name}{character: path of surface path .ascii extension file}

\item{npoint}{numeric: number of evenly spaced points desidered}

\item{mag}{numeric: how many times will be divided by the number of initial points, if NULL dec.curve will be suppress}
}
\value{
evenly.set numeric: a kxd matrix with evenly-spaced landmark coordinates
}
\description{
This function find npoints evenly spaced from a surface path (ascii format) from Amira
}
\author{
Antonio Profico, Alessio Veneziano
}

