% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bezier.amira.set.r
\name{bezier.amira.set}
\alias{bezier.amira.set}
\title{bezier.amira.set}
\usage{
bezier.amira.set(name.file, nland, npoint, mag = 3)
}
\arguments{
\item{name.file}{character: path of landmark file set (.landmarkAscii)}

\item{npoint}{numeric: number of evenly spaced points desidered}

\item{mag}{numeric: how many times will be divided by the number of initial points, if NULL dec.curve will be suppress}
}
\value{
evenly.set numeric: matrix with evenly-spaced landmark coordinates (kxd)
}
\description{
This function find npoints evenly spaced from a landmark file set (.landmarkAscii)
}
\author{
Antonio Profico, Alessio Veneziano
}

