% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dec.curve.r
\name{dec.curve}
\alias{dec.curve}
\title{dec.curve}
\usage{
dec.curve(mat_input, mag, plot = TRUE)
}
\arguments{
\item{mat_input}{numeric matrix}

\item{mag}{numeric: how many times will be divided by the number of initial points}

\item{plot}{logical: if TRUE will be plotted the starting and final point matrices}
}
\value{
matt numeric: a kxd matrix with landmark coordinates
}
\description{
This function find and order a series of intermediate points
}
\author{
Antonio Profico, Alessio Veneziano
}

