% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_amira.r
\name{export_amira}
\alias{export_amira}
\title{export_amira}
\usage{
export_amira(lista, path = getwd())
}
\arguments{
\item{path}{character: path of the folder where saving the Amira landmark sets}

\item{lista:}{list of 3D landmark sets}
}
\description{
This function export a list of 3D landmark set in separate files (format landmarkAscii)
}
\author{
Antonio Profico, Alessio Veneziano
}

