% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imagematrix.R
\name{imagematrix}
\alias{imagematrix}
\title{imagematrix}
\usage{
imagematrix(mat, type = NULL, ncol = dim(mat)[1], nrow = dim(mat)[2],
  noclipping = FALSE)
}
\description{
Function from the removed package rimage
}
\author{
Tomomi TAKASHINA
}

