% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iter.one2four.r
\name{iter.one2four}
\alias{iter.one2four}
\title{iter.one2four}
\usage{
\method{iter}{one2four}(mesh, iter)
}
\arguments{
\item{mesh}{triangular mesh stored as object of class "mesh3d"}

\item{iter}{numeric; number of iteration}
}
\value{
prova list with 4^(iter+1) separated mesh
}
\description{
This repeat iteratively the function one2four
}
\author{
Antonio Profico
}

