% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.amira.dir.r
\name{list.amira.dir}
\alias{list.amira.dir}
\title{list.amira.dir}
\usage{
list.amira.dir(path.dir, nland)
}
\arguments{
\item{path.dir:}{path directory where Amira landmark set are stored}

\item{nland:}{number of landmark, if set on "auto" the value will be automatically set}
}
\value{
list.amira numeric: a list with landmark coordinates (kxdxn)
}
\description{
This function create a list formed by files of landmark set of Amira stored in a folder
}
\author{
Antonio Profico, Alessio Venezianp
}

