% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/noise.mesh.R
\name{noise.mesh}
\alias{noise.mesh}
\title{noise.mesh}
\usage{
noise.mesh(mesh, noise = 0.025, seed = 123)
}
\arguments{
\item{mesh}{triangular mesh stored as object of class "mesh3d"}

\item{noise}{sd deviation to defin vertex shifting}

\item{seed}{seed for random number generator}
}
\value{
mesh_n mesh noised
}
\description{
This function add noise to a mesh
}
\author{
Antonio Profico
}

