% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/one2four.r
\name{one2four}
\alias{one2four}
\title{one2four}
\usage{
one2four(mesh, plot = FALSE)
}
\arguments{
\item{mesh}{triangular mesh stored as object of class "mesh3d"}

\item{logical:}{if TRUE plot the final 4 separeted mesh}
}
\value{
storage.mesh list with 4 separated mesh
}
\description{
This function divide a mesh in 4 parts
}
\author{
Antonio Profico
}

