% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order.open.outline.r
\name{order.open.outline}
\alias{order.open.outline}
\title{order.open.outline}
\usage{
order.open.outline(matrix, to.2D = FALSE)
}
\arguments{
\item{matrix}{numeric: matrix of outline coordinates (3D)}

\item{to.2D}{logical: if TRUE the ordered matrix will be converted in 2D}
}
\value{
matt numeric: matrix with ordered points
}
\description{
This function order an outline along the xz plane and eventually convert it in 2 dimensions
}
\author{
Antonio Profico, Alessio Veneziano
}

