% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.amira.dir.r
\name{read.amira.dir}
\alias{read.amira.dir}
\title{read.amira.dir}
\usage{
read.amira.dir(path.dir, nland)
}
\arguments{
\item{path.dir}{character: path of the folder}

\item{nland}{numeric: number of landmark sampled in Amira}
}
\value{
array.set numeric: a kxdxn array with landmark coordinates
}
\description{
This function read and store in array the coordinated allocated in a folder in separate files (format
}
\author{
Antonio Profico, Alessio Veneziano
}

