% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.amira.set.r
\name{read.amira.set}
\alias{read.amira.set}
\title{read.amira.set}
\usage{
read.amira.set(name.file, nland)
}
\arguments{
\item{name.file}{character: path of landamrkascii file}

\item{nland}{numeric: number of landmark sampled in Amira, if is set on "auto" will be automatically recognized the landmark number}
}
\value{
array.set numeric: an array with landmark coordinates (kxdxn)
}
\description{
This function convert a file landmark set of Amira in a array
}
\author{
Antonio Profico, Alessio Veneziano
}

